Moved wsgi components to their proper module.
Refs #1
