+++ nutrac/app.py
@@ -1,129 +1,35 @@
+#!/usr/bin/env python
+#
+# Copyright 2015-2018 Flavio Garcia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from . import handlers
-from firenado import tornadoweb, session
-import tornado.web
-
-import base64
+from . import wsgi
+from firenado import tornadoweb
 import os
-import sys
-
-import logging
-import logging.handlers
-
-import tornado.web
-import tornado.ioloop
-import tornado.httpserver
-import tornado.options
-import tornado.wsgi
-from tornado import escape
-from tornado import httputil
-
-import trac.web.main
-
-
-def application(environ, start_response, component, handler, request):
-    trac_root = component.conf['trac']['root']
-    project_relative = "/".join(request.uri.split("/")[1:3])
-    project_path = os.path.join(trac_root, project_relative)
-    os.environ['TRAC_ENV'] = project_path
-    environ['PATH_INFO'] = environ['PATH_INFO'].replace("/%s" %
-
-                                                        project_relative, "")
-    if component.project_exists(project_relative):
-        environ['SCRIPT_NAME'] = "/%s/" % project_relative
-        #environ['REMOTE_USER'] = "buga"
-
-        request.application = component.application
-        #environ['trac.env_path'] = os.path.join(PROJECT_ROOT, '..', '..')
-        #environ['trac.base_path'] = 'candango'
-        #if 'HTTP_AUTHORIZATION' in environ:
-            #auth_header = environ['HTTP_AUTHORIZATION']
-            #environ['REMOTE_USER'] = base64.decodestring(auth_header[6:]).split(':')[0]
-        #environ['SCRIPT_NAME'] = os.path.join('/', sys.argv[1])
-        return trac.web.main.dispatch_request(environ, start_response)
-    else:
-        status = "404 Not Found"
-        response_headers = [("Content-type", "text/plain")]
-        start_response(status, response_headers)
-        return ["File not found"]
-
-
-class ComponentizedWSGIContainer(tornado.wsgi.WSGIContainer):
-
-    def __init__(self, wsgi_application, component):
-        super(ComponentizedWSGIContainer, self).__init__(wsgi_application)
-        self.component = component
-        self.handler = None
-
-    def __call__(self, request):
-        data = {}
-        response = []
-
-        def start_response(status, response_headers, exc_info=None):
-            data["status"] = status
-            data["headers"] = response_headers
-            return response.append
-        app_response = self.wsgi_application(
-            ComponentizedWSGIContainer.environ(request), start_response,
-            self.component, self.handler, request)
-        self.handler = None
-        try:
-            response.extend(app_response)
-            body = b"".join(response)
-        finally:
-            if hasattr(app_response, "close"):
-                app_response.close()
-        if not data:
-            raise Exception("WSGI app did not call start_response")
-
-        status_code, reason = data["status"].split(' ', 1)
-        status_code = int(status_code)
-        headers = data["headers"]
-        header_set = set(k.lower() for (k, v) in headers)
-        body = escape.utf8(body)
-        if status_code != 304:
-            if "content-length" not in header_set:
-                headers.append(("Content-Length", str(len(body))))
-            if "content-type" not in header_set:
-                headers.append(("Content-Type", "text/html; charset=UTF-8"))
-        if "server" not in header_set:
-            headers.append(("Server", "TornadoServer/%s" % tornado.version))
-
-        start_line = httputil.ResponseStartLine("HTTP/1.1", status_code, reason)
-        header_obj = httputil.HTTPHeaders()
-        for key, value in headers:
-            header_obj.add(key, value)
-        request.connection.write_headers(start_line, header_obj, chunk=body)
-        request.connection.finish()
-        self._log(status_code, request)
-
-
-class ComponentizedFallbackHandler(tornado.web.FallbackHandler):
-
-    def initialize(self, component, fallback):
-        self.component = component
-        fallback.handler = self
-        super(ComponentizedFallbackHandler, self).initialize(fallback)
-
-    @session.read
-    def prepare(self):
-        super(ComponentizedFallbackHandler, self).prepare()
-
-    @session.write
-    def on_finish(self):
-        pass
-        # self.component.run_after_handler(self)
 
 
 class NutracComponent(tornadoweb.TornadoComponent):
 
     def get_handlers(self):
-        container = ComponentizedWSGIContainer(application, self)
+        container = wsgi.ComponentizedWSGIContainer(wsgi.application, self)
         return [
             (r"/", handlers.IndexHandler),
             (r"/profile", handlers.ProfileHandler),
             (r"/login", handlers.ProfileHandler),
             (r"/([\w|\-|\_|\@|]*\/?)", handlers.HomeHandler),
-            (r"/([\w|\-|\_|\@|]*)/.*", ComponentizedFallbackHandler,
+            (r"/([\w|\-|\_|\@|]*)/.*", wsgi.ComponentizedFallbackHandler,
              dict(component=self, fallback=container))
         ]
